requirements.txt
asgiref==3.8.1
branca==0.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
folium==0.17.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
openrouteservice==2.3.3
pandas==2.2.2
polyline==2.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
xyzservices==2024.6.0
